+++ src/api/RuleTypes.ts
@@ -411,6 +411,14 @@ export interface IVarRule<K extends VarTemplateName = any> extends INamedEntity,
      */
 	readonly template: K;
 
+	/** Custom CSS property name prefixed with `"--"` */
+	readonly cssVarName: string;
+
+    /**
+	 * Gets the value of the property.
+	 */
+	getValue(): ExtendedVarValue<K>;
+
     /**
 	 * Sets new value of this custom CSS property at the global level; that is, under `:root`. To
      * set a value of the CSS custom property under a certain CSS rule, use the
@@ -629,7 +637,7 @@ export interface IGroupRule<T extends IStyleDefinition = any> extends IRule
 	readonly condition: string;
 
 	// Instance of the style definition class defining the rules under this grouping rule
-	readonly rules: T;
+	readonly definition: T;
 
 	/** CSSOM grouping rule */
 	readonly cssRule: CSSGroupingRule | null;