+++ src/api/StyleTypes.ts
@@ -259,7 +259,7 @@ export type BackfaceVisibilityMode_StyleType = "visible" | "hidden";
  * Type for single background value
  * @category Style Helper
  */
-export type Background_Single = CssColor | CssImage |
+export type Background_Single = string | CssColor | CssImage |
     {
         color?: Extended<CssColor>,
         image?: Extended<CssImage>,