+++ src/test/specs/testStyleRules.ts
@@ -25,12 +25,12 @@ describe("style rules:", () =>
 		}
 
 		let a = css.activate( A);
-		expect(a!.cls.cssRule).not.toBeDefined();
-		expect(a!.id.cssRule).not.toBeDefined();
-		expect(a!.tag.cssRule).not.toBeDefined();
-		expect(a!.selector.cssRule).not.toBeDefined();
+		expect(a.cls.cssRule).not.toBeDefined();
+		expect(a.id.cssRule).not.toBeDefined();
+		expect(a.tag.cssRule).not.toBeDefined();
+		expect(a.selector.cssRule).not.toBeDefined();
 
-		css.deactivate( a!);
+		css.deactivate( a);
 	})
 
 
@@ -46,16 +46,16 @@ describe("style rules:", () =>
 		}
 
 		let a = css.activate( A);
-		expect(a!.cls.cssRule).toBeTruthy();
-		expect(a!.id.cssRule).toBeTruthy();
-		expect(a!.tag.cssRule).toBeTruthy();
-		expect(a!.selector.cssRule).toBeTruthy();
+		expect(a.cls.cssRule).toBeTruthy();
+		expect(a.id.cssRule).toBeTruthy();
+		expect(a.tag.cssRule).toBeTruthy();
+		expect(a.selector.cssRule).toBeTruthy();
 
-		css.deactivate( a!);
-		expect(a!.cls.cssRule).toBeFalsy();
-		expect(a!.id.cssRule).toBeFalsy();
-		expect(a!.tag.cssRule).toBeFalsy();
-		expect(a!.selector.cssRule).toBeFalsy();
+		css.deactivate( a);
+		expect(a.cls.cssRule).toBeFalsy();
+		expect(a.id.cssRule).toBeFalsy();
+		expect(a.tag.cssRule).toBeFalsy();
+		expect(a.selector.cssRule).toBeFalsy();
 	})
 
 
@@ -71,12 +71,12 @@ describe("style rules:", () =>
 		}
 
 		let a = css.activate( A);
-		expect(a!.cls.cssRule?.selectorText).toEqual(".A_cls");
-		expect(a!.id.cssRule?.selectorText).toEqual("#A_id");
-		expect(a!.tag.cssRule?.selectorText).toEqual("h1");
-		expect(a!.selector.cssRule?.selectorText).toEqual("div > span");
+		expect(a.cls.cssRule?.selectorText).toEqual(".A_cls");
+		expect(a.id.cssRule?.selectorText).toEqual("#A_id");
+		expect(a.tag.cssRule?.selectorText).toEqual("h1");
+		expect(a.selector.cssRule?.selectorText).toEqual("div > span");
 
-		css.deactivate( a!);
+		css.deactivate( a);
 	})
 
 
@@ -104,7 +104,7 @@ describe("style rules:", () =>
 		expect((cssRules[1] as CSSStyleRule).selectorText).toEqual(".A_cls:hover");
 		expect((cssRules[2] as CSSStyleRule).selectorText).toEqual(".A_cls:hover:visited");
 
-		css.deactivate( a!);
+		css.deactivate( a);
 	})
 
 
@@ -129,7 +129,7 @@ describe("style rules:", () =>
 		expect((cssRules[0] as CSSStyleRule).selectorText).toEqual(".A_cls2 > .A_cls1");
 		expect((cssRules[1] as CSSStyleRule).selectorText).toEqual(".A_cls1 + .A_cls2");
 
-		css.deactivate( a!);
+		css.deactivate( a);
 	})
 
 
@@ -160,7 +160,7 @@ describe("style rules:", () =>
 		expect((cssRules[2] as CSSStyleRule).selectorText).toEqual(".A_cls.other, .other.A_cls div > span");
 		expect((cssRules[3] as CSSStyleRule).selectorText).toEqual(".A_cls.A_cls1, .A_cls1.A_cls + table");
 
-		css.deactivate( a!);
+		css.deactivate( a);
 	})
 
 
@@ -194,7 +194,7 @@ describe("style rules:", () =>
 		expect((cssRules[4] as CSSStyleRule).selectorText).toEqual(".A_clsCopy.other, .other.A_clsCopy div > span");
 		expect((cssRules[5] as CSSStyleRule).selectorText).toEqual(".A_clsCopy > table");
 
-		css.deactivate( a!);
+		css.deactivate( a);
 	})