+++ src/test/utils/dom.ts
@@ -84,9 +84,9 @@ export function testLonghandProp<K extends keyof Styleset>( propName: K, propVal
 	}
 
 	let a = css.activate( A);
-	verifyPropValue( a!.c, propName, ...expected);
+	verifyPropValue( a.c, propName, ...expected);
 
-	css.deactivate( a!);
+	css.deactivate( a);
 }
 
 
@@ -101,9 +101,9 @@ export function testShorthandProp<K extends keyof Styleset>( propName: K, propVa
 	}
 
 	let a = css.activate( A);
-	verifyMultiPropValues( a!.c, expected);
+	verifyMultiPropValues( a.c, expected);
 
-	css.deactivate( a!);
+	css.deactivate( a);
 }
 
 
@@ -117,9 +117,9 @@ export function testPropPriority<K extends keyof Styleset>( propName: K, propVal
 	}
 
 	let a = css.activate( A);
-	verifyPropPriority( a!.c, propName, expectedPriority);
+	verifyPropPriority( a.c, propName, expectedPriority);
 
-	css.deactivate( a!);
+	css.deactivate( a);
 }